server/requirements.txt
# Web Framework
fastapi
uvicorn[standard]
python-multipart

# LangChain Core
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-groq

# Vector Store
pinecone

# Embeddings (Local HuggingFace)
langchain-huggingface
sentence-transformers
transformers
torch

# PDF Parsing
pypdf

# Environment Variables
python-dotenv

# Utilities
pydantic
requests
tqdm

# Logging
loguru
